+++ NEWS.md
@@ -1,7 +1,14 @@
+# ggsurveillance 0.2.0
+
+-   Update geom_epigantt(): Add auto-scaling for linewidth and update documentation
+-   Add scale_y_discrete_reverse()
+-   New dataset of a fictional hospital outbreak
+-   geometric_mean(): Add an option to disable warnings
+-   Minor fixes
+
 # ggsurveillance 0.1.2
 
 -   Documentation improvements and bug fixes for geom_epicurve()
-
 
 # ggsurveillance 0.1.1